Feature/product (#143)
* feat(product): 기본적인 파일 생성

* 임시저장

* 임시저장2

* feat(ProductServiceImpl): 기능 구현 완료

* feat(ProductServiceImpl): 기능 구현 완료2

* modify(ProductServiceImpl): Exception 오류 수정

* modify(Product, ProductServiceImpl): 생성자DTO로 입력하도록 수정

* service 구현완료

* 진헹중

* feat(Product): Product에서 AdminController와 테스트코드 구현완료

* modify(TagServiceImpl, ProductServiceImpl): @Transactional 추가

* modify(RequestProductCreateDTO): 필드명 수정

* modify(RequestDTO): ProductStateId, publisherId 아이디를 네임으로 받게 변경

* 임시저장

* 최신화

* feat(Product): api 사용하여 도서 등록 진행중

* feat(Product): api 사용하여 도서 등록 진행중

* feat(Product): api 사용하여 도서 등록 진행중

* feat(Product): 도서검색시 페이징 적용 추가

* feat(Product): 진행중

* product-controbutor 임시저장

* feat(Product): 진행중

* feat: 관게테이블 저장하는 코드 추가

* product 수정 전 저장(name -> id)

* product 수정 전 저장(name -> id).

* feat(Product): api 이용해서 등록 기능 구현 완료

* feat(Product): 테스트

* feat(Product): 테스트

* modify(Product): 오류 수정

* modify(Product): 메소드 타입, 이름 수정

* modify(Product): 도서등록시 카테고리, 태크 추가

* feature(product): 진행중

* feature(product): 기능 구현 완료

* feature(product) : api 사용 등록 기능 구현 완료

* ..

* modify(product) : 이름 수정

* feat(product) : bestseller, 신간도서 불러오기 기능 추가

* feat(product-mainPage) : 메인페이지에서 카테고리별로 책 보여주기

* feat(product-mainPage) : 경로 수정

* feat(product-mainPage) : 메인화면에 책 불러오기 구현

* feat(product-mainPage) : 더미데이터 추가

* modify(contributor, position, tag): @Admin, @Transactional 어노테이션 붙임

* modify(product): @Transactional 처리 및 메서드 이름 수정

* 오버라이드 어노테이션 수정

* modify(sample.sql): review 잘못 삭제해서 추가

* 파일 최종병합

* modify(apiProduct, tag): tag 미선택 시 등록 가능하게 수정

* feat(product): 도서 직접 등록 최종 구현

* 도서 수정 및 삭제 완료

* 누락된 import 추가

* product테스트 주석

---------

+++ src/main/java/com/nhnacademy/back/product/product/service/impl/ProductServiceImpl.java
@@ -1,17 +1,19 @@
 package com.nhnacademy.back.product.product.service.impl;
 
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
+import java.util.Objects;
+import java.util.UUID;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
+import org.thymeleaf.util.StringUtils;
 
-import com.nhnacademy.back.product.category.domain.dto.response.ResponseCategoryDTO;
-import com.nhnacademy.back.product.category.domain.entity.Category;
+import com.nhnacademy.back.common.util.MinioUtils;
 import com.nhnacademy.back.product.category.repository.ProductCategoryJpaRepository;
-import com.nhnacademy.back.product.contributor.domain.dto.response.ResponseContributorDTO;
 import com.nhnacademy.back.product.contributor.domain.entity.Contributor;
 import com.nhnacademy.back.product.contributor.domain.entity.ProductContributor;
 import com.nhnacademy.back.product.contributor.exception.ContributorNotFoundException;
@@ -40,7 +42,6 @@ import com.nhnacademy.back.product.state.domain.entity.ProductState;
 import com.nhnacademy.back.product.state.domain.entity.ProductStateName;
 import com.nhnacademy.back.product.state.exception.ProductStateNotFoundException;
 import com.nhnacademy.back.product.state.repository.ProductStateJpaRepository;
-import com.nhnacademy.back.product.tag.domain.dto.response.ResponseTagDTO;
 import com.nhnacademy.back.product.tag.domain.entity.ProductTag;
 import com.nhnacademy.back.product.tag.domain.entity.Tag;
 import com.nhnacademy.back.product.tag.exception.TagNotFoundException;
@@ -63,6 +64,10 @@ public class ProductServiceImpl implements ProductService {
 	private final TagJpaRepository tagJpaRepository;
 	private final ProductTagJpaRepository productTagJpaRepository;
 
+	private final MinioUtils minioUtils;
+	private final String BUCKET_NAME = "e2bs-products-image";
+
+
 	/**
 	 * 도서를 DB에 저장
 	 * 도서가 이미 존재하면 Exception 발생
@@ -75,13 +80,19 @@ public class ProductServiceImpl implements ProductService {
 			throw new ProductAlreadyExistsException("Product already exists");
 		}
 
-		ProductState productState = productStateJpaRepository.findById(request.getProductStateId())
-			.orElseThrow(() -> new ProductStateNotFoundException("상품상태 조회 실패"));
+		//상태 저장
+		ProductState productState = productStateJpaRepository.findByProductStateId(request.getProductStateId());
+		Integer productStock = request.getProductStock();
+		ProductState stateSale = productStateJpaRepository.findByProductStateName(ProductStateName.SALE);
+
+		if (productStock == 0 && productState == stateSale) {
+			productState = productStateJpaRepository.findByProductStateName(ProductStateName.OUT);
+		}
 
 		Publisher publisher = publisherJpaRepository.findById(request.getPublisherId())
 			.orElseThrow(() -> new PublisherNotFoundException("출판사 조회 실패"));
 
-		List<String> imagePaths = request.getProductImagePaths();
+		List<MultipartFile> productImageFiles = request.getProductImage();
 		List<Long> tagIds = request.getTagIds();
 		List<Long> contributorIds = request.getContributorIds();
 
@@ -93,17 +104,23 @@ public class ProductServiceImpl implements ProductService {
 
 		// 이미지 저장
 		// 자식 추가 (ProductImage)
-		for (String imagePath : imagePaths) {
+		for (MultipartFile productImageFile : productImageFiles) {
+			String imagePath = "";
+			if (Objects.nonNull(productImageFile) && !productImageFile.isEmpty()) {
+				imagePath = uploadFile(productImageFile);
+			}
 			ProductImage productImage = new ProductImage(product, imagePath);
 			product.getProductImage().add(productImage);
 			productImageJpaRepository.save(productImage);
 		}
 
 		// 태그 저장
-		for (Long tagId : tagIds) {
-			Tag tag = tagJpaRepository.findById(tagId)
-				.orElseThrow(() -> new TagNotFoundException("태그 조회 실패"));
-			productTagJpaRepository.save(new ProductTag(product, tag));
+		if (!Objects.isNull(request.getTagIds())) {
+			for (Long tagId : tagIds) {
+				Tag tag = tagJpaRepository.findById(tagId)
+					.orElseThrow(() -> new TagNotFoundException("태그 조회 실패"));
+				productTagJpaRepository.save(new ProductTag(product, tag));
+			}
 		}
 
 		// 기여자 저장
@@ -123,27 +140,7 @@ public class ProductServiceImpl implements ProductService {
 	public ResponseProductReadDTO getProduct(long productId) {
 		Product product = productJpaRepository.findById(productId)
 			.orElseThrow(ProductNotFoundException::new);
-
-		return new ResponseProductReadDTO(
-			product.getProductId(),
-			new ResponseProductStateDTO(product.getProductState().getProductStateId(),
-				product.getProductState().getProductStateName().name()),
-			new ResponsePublisherDTO(product.getPublisher().getPublisherId(),
-				product.getPublisher().getPublisherName()),
-			product.getProductTitle(),
-			product.getProductContent(),
-			product.getProductDescription(),
-			product.getProductPublishedAt(),
-			product.getProductIsbn(),
-			product.getProductRegularPrice(),
-			product.getProductSalePrice(),
-			product.isProductPackageable(),
-			product.getProductStock(),
-			productImageJpaRepository.findImageDTOsByProductId(productId),
-			productTagJpaRepository.findTagDTOsByProductId(productId),
-			productCategoryJpaRepository.findCategoryDTOsByProductId(productId),
-			productContributorJpaRepository.findContributorDTOsByProductId(productId)
-		);
+		return getProductByChangedImagePath(product);
 	}
 
 	/**
@@ -158,51 +155,8 @@ public class ProductServiceImpl implements ProductService {
 		} else {
 			productPage = productJpaRepository.findAllByCategoryId(categoryId, pageable);
 		}
-		List<Long> productIds = productPage.stream().map(Product::getProductId).toList();
 
-		Map<Long, List<ProductImage>> imageMap = productImageJpaRepository.findAllByProductIdsGrouped(productIds);
-		Map<Long, List<Tag>> tagMap = productTagJpaRepository.findTagsGroupedByProductIds(productIds);
-		Map<Long, List<Category>> categoryMap = productCategoryJpaRepository.findCategoriesGroupedByProductIds(
-			productIds);
-		Map<Long, List<Contributor>> contributorMap = productContributorJpaRepository.findContributorsGroupedByProductIds(
-			productIds);
-
-		return productPage.map(product -> {
-			Long id = product.getProductId();
-			return new ResponseProductReadDTO(
-				id,
-				new ResponseProductStateDTO(product.getProductState().getProductStateId(),
-					product.getProductState().getProductStateName().name()),
-				new ResponsePublisherDTO(product.getPublisher().getPublisherId(),
-					product.getPublisher().getPublisherName()),
-				product.getProductTitle(),
-				product.getProductContent(),
-				product.getProductDescription(),
-				product.getProductPublishedAt(),
-				product.getProductIsbn(),
-				product.getProductRegularPrice(),
-				product.getProductSalePrice(),
-				product.isProductPackageable(),
-				product.getProductStock(),
-				imageMap.getOrDefault(id, List.of())
-					.stream()
-					.map(image -> new ResponseProductImageDTO(image.getProductImageId(), image.getProductImagePath()))
-					.toList(),
-				tagMap.getOrDefault(id, List.of()).stream()
-					.map(tag -> new ResponseTagDTO(tag.getTagId(), tag.getTagName())).toList(),
-				categoryMap.getOrDefault(id, List.of())
-					.stream()
-					.map(
-						category -> new ResponseCategoryDTO(category.getCategoryId(), category.getCategoryName(), null))
-					.toList(),
-				contributorMap.getOrDefault(id, List.of())
-					.stream()
-					.map(contributor -> new ResponseContributorDTO(contributor.getContributorId(),
-						contributor.getContributorName(),
-						contributor.getPosition().getPositionId(), contributor.getPosition().getPositionName()))
-					.toList()
-			);
-		});
+		return productPage.map(this::getProductByChangedImagePath);
 	}
 
 	/**
@@ -218,26 +172,7 @@ public class ProductServiceImpl implements ProductService {
 		}
 
 		return products.stream()
-			.map(product -> new ResponseProductReadDTO(
-				product.getProductId(),
-				new ResponseProductStateDTO(product.getProductState().getProductStateId(),
-					product.getProductState().getProductStateName().name()),
-				new ResponsePublisherDTO(product.getPublisher().getPublisherId(),
-					product.getPublisher().getPublisherName()),
-				product.getProductTitle(),
-				product.getProductContent(),
-				product.getProductDescription(),
-				product.getProductPublishedAt(),
-				product.getProductIsbn(),
-				product.getProductRegularPrice(),
-				product.getProductSalePrice(),
-				product.isProductPackageable(),
-				product.getProductStock(),
-				productImageJpaRepository.findImageDTOsByProductId(product.getProductId()),
-				productTagJpaRepository.findTagDTOsByProductId(product.getProductId()),
-				productCategoryJpaRepository.findCategoryDTOsByProductId(product.getProductId()),
-				productContributorJpaRepository.findContributorDTOsByProductId(product.getProductId())
-			)).toList();
+			.map(this::getProductByChangedImagePath).toList();
 	}
 
 	/**
@@ -255,28 +190,82 @@ public class ProductServiceImpl implements ProductService {
 		Publisher publisher = publisherJpaRepository.findById(request.getPublisherId())
 			.orElseThrow(() -> new PublisherNotFoundException("출판사 조회 실패"));
 
-		// 상품 정보 업데이트 (수정)
-		product.updateProduct(request, productState, publisher);
 
-		List<String> imagePaths = request.getProductImagePaths();
+
+		List<MultipartFile> productImageFiles = request.getProductImage();
 		List<Long> tagIds = request.getTagIds();
 		List<Long> contributorIds = request.getContributorIds();
 
-		// 이미지 삭제 후 저장
+
+
 		// 자식 추가 (ProductImage)
-		productImageJpaRepository.deleteByProduct_ProductId(productId);
-		for (String imagePath : imagePaths) {
-			ProductImage productImage = new ProductImage(product, imagePath);
-			product.getProductImage().add(productImage);
-			productImageJpaRepository.save(productImage);
+		// - 이미지가 들어왔다면
+		if (Objects.nonNull(productImageFiles) && !productImageFiles.isEmpty()) {
+			// - 기존 리스트 조회
+			List<ProductImage> productImages = productImageJpaRepository.getAllByProduct_ProductId(productId);
+
+			// - miniO 기존 파일 삭제
+			for (ProductImage productImage : product.getProductImage()) {
+				minioUtils.deleteObject(BUCKET_NAME, productImage.getProductImagePath());
+			}
+
+			// - DB에서 삭제
+			productImageJpaRepository.deleteAll(productImages);
+
+
+			List<ProductImage> newProductImages = new ArrayList<>();
+			for (MultipartFile productImageFile : productImageFiles) {
+				String originalFilename = productImageFile.getOriginalFilename();
+				String objectName = UUID.randomUUID() + "_" + originalFilename;
+
+				// MinIO에 파일 업로드
+				minioUtils.uploadObject(BUCKET_NAME, objectName, productImageFile);
+
+				// 새로운 ProductImage 엔티티 생성
+				ProductImage productImage = new ProductImage(product, objectName);
+				newProductImages.add(productImage);
+			}
+			// Product의 productImage 리스트 갱신
+			product.getProductImage().clear(); // 기존 리스트 비우기
+			product.getProductImage().addAll(newProductImages);
+			productImageJpaRepository.saveAll(newProductImages); // 새로운 이미지 저장
 		}
 
+		// 	// - 수정해서 받은 사진의 갯수만큼 반복
+		// 	for (MultipartFile productImageFile : productImageFiles) {
+		//
+		//
+		// 		// - miniO에 새로 업로드할 파일이름
+		// 		String originalFilename = productImageFile.getOriginalFilename();
+		//
+		// 		// - 빈 경로 생성
+		// 		String updateImagePath = "";
+		//
+		// 		// - 이름 설정 후 업로드
+		// 		UUID uuid = UUID.randomUUID();
+		// 		String objectName = uuid + "_" + originalFilename;
+		// 		minioUtils.uploadObject(BUCKET_NAME, objectName, productImageFile);
+		//
+		// 		// 가공된 파일명 적용
+		// 		updateImagePath = objectName;
+		//
+		// 		ProductImage productImage = new ProductImage(product, updateImagePath);
+		// 		product.getProductImage().add(productImage);
+		// 		productImageJpaRepository.save(productImage);
+		// 	}
+		// }
+
+		// 상품 정보 업데이트 (수정)
+		product.updateProduct(request, productState, publisher);
+
 		// 태그 삭제 후 저장
-		productTagJpaRepository.deleteByProduct_ProductId(productId);
-		for (Long tagId : tagIds) {
-			Tag tag = tagJpaRepository.findById(tagId)
-				.orElseThrow(() -> new TagNotFoundException("태그 조회 실패"));
-			productTagJpaRepository.save(new ProductTag(product, tag));
+		if (!Objects.isNull(request.getTagIds())) {
+			productTagJpaRepository.deleteByProduct_ProductId(productId);
+			for (Long tagId : tagIds) {
+				Tag tag = tagJpaRepository.findById(tagId)
+					.orElseThrow(() -> new TagNotFoundException("태그 조회 실패"));
+				productTagJpaRepository.save(new ProductTag(product, tag));
+			}
 		}
 
 		// 기여자 삭제 후 저장
@@ -335,6 +324,56 @@ public class ProductServiceImpl implements ProductService {
 				product.getProductTitle(),
 				product.getPublisher().getPublisherName()
 			));
+	}
+
+	/**
+	 * 파일 업로드 메소드
+	 */
+	private String uploadFile(MultipartFile productImageFile) {
+		String originalFilename = productImageFile.getOriginalFilename();
+		UUID uuid = UUID.randomUUID();
+		String objectName = uuid + "_" + originalFilename;
+		minioUtils.uploadObject(BUCKET_NAME, objectName, productImageFile);
+		return objectName;
+	}
+
+	/**
+	 * 이미지 경로 가공 메소드
+	 */
+	private ResponseProductReadDTO getProductByChangedImagePath(Product product) {
+		String productImagePath = "";
+		List<String> productImagePaths = productImageJpaRepository.findAllByProduct_ProductId(product.getProductId());
+		List<ResponseProductImageDTO> changedResponseProductImageDTOs = new ArrayList<>();
+
+		for (String productImage : productImagePaths){
+			if (!StringUtils.isEmpty(productImage)) {
+				productImagePath = minioUtils.getPresignedUrl(BUCKET_NAME, productImage);
+				ResponseProductImageDTO responseProductImageDTO = new ResponseProductImageDTO(productImageJpaRepository.findByProductImagePath(productImage), productImagePath);
+				changedResponseProductImageDTOs.add(responseProductImageDTO);
+			}
+		}
+
+		return new ResponseProductReadDTO(
+			product.getProductId(),
+			new ResponseProductStateDTO(product.getProductState().getProductStateId(),
+				product.getProductState().getProductStateName().name()),
+			new ResponsePublisherDTO(product.getPublisher().getPublisherId(),
+				product.getPublisher().getPublisherName()),
+			product.getProductTitle(),
+			product.getProductContent(),
+			product.getProductDescription(),
+			product.getProductPublishedAt(),
+			product.getProductIsbn(),
+			product.getProductRegularPrice(),
+			product.getProductSalePrice(),
+			product.isProductPackageable(),
+			product.getProductStock(),
+			changedResponseProductImageDTOs,
+			productTagJpaRepository.findTagDTOsByProductId(product.getProductId()),
+			productCategoryJpaRepository.findCategoryDTOsByProductId(product.getProductId()),
+			productContributorJpaRepository.findContributorDTOsByProductId(product.getProductId()
+			)
+		);
 	}
 
 }
